requirements.txt
patool==1.15.0
Send2Trash==1.8.2
watchdog==3.0.0
